Add files via upload
A deep learning model for predicting house prices, trained using a Sequential neural network with K-Fold cross-validation. The model features dense layers with ReLU activation, dropout for regularization, and is optimized with the Adam optimizer. It has been trained on scaled data and is saved for use in predicting housing prices on test datasets.